chore: move index.js to client/index.js
