+++ tests/integration/instance.js
@@ -1,7 +1,7 @@
 var store = require('humble-localstorage')
 var test = require('tape')
 
-var ConnectionStatus = require('../../index')
+var ConnectionStatus = require('../../client')
 
 test('Instance', function (t) {
   var connectionStatus = new ConnectionStatus('https://example.com/ping')
